update to net6.0

+++ src/AmberEggApi.Integration.Tests/Factories/PersonaControllerFactoryTest.cs
@@ -1,13 +1,9 @@
 ﻿using AmberEggApi.ApplicationService.ViewModels;
 using AmberEggApi.Domain.Commands;
-using Api.Common.WebServer.Server;
 using FluentAssertions;
 using Newtonsoft.Json;
-using System;
 using System.Net;
-using System.Net.Http;
 using System.Text;
-using System.Threading.Tasks;
 
 namespace AmberEggApi.Integration.Tests.Factories
 {
@@ -26,14 +22,9 @@ namespace AmberEggApi.Integration.Tests.Factories
             var name = "Persona Test";
 
             //Act
-            // var responseModel = await Create(new CreatePersonaCommand(name));
-            // var viewModel =
-            //     JsonConvert.DeserializeObject<PersonaViewModel>(responseModel.Result.ToString());
-
-             var viewModel = await Create(new CreatePersonaCommand(name));
+            var viewModel = await Create(new CreatePersonaCommand(name));
 
             // Assert
-            //responseModel.StatusCode.Should().Be((int)HttpStatusCode.Created);
             viewModel.Should().BeOfType<PersonaViewModel>();
 
             viewModel.Id.Should().NotBeEmpty();
@@ -46,12 +37,9 @@ namespace AmberEggApi.Integration.Tests.Factories
         {
             // Act
             var response = await client.DeleteAsync($"{url}/{id}");
-            //var responseModel = JsonConvert.DeserializeObject<ApiResponse>(await response.Content.ReadAsStringAsync());
 
             // Assert
             response.StatusCode.Should().Be(HttpStatusCode.NoContent);
-            //responseModel.StatusCode.Should().Be((int)HttpStatusCode.NoContent);
-            //responseModel.Result.Should().Be("");
         }
 
         public async Task<PersonaViewModel> Get(Guid id)
@@ -60,7 +48,7 @@ namespace AmberEggApi.Integration.Tests.Factories
             // Act
             var response = await client.GetAsync($"{url}/{id}");
 
-            if(response.StatusCode == HttpStatusCode.NoContent)
+            if (response.StatusCode == HttpStatusCode.NoContent)
             {
                 return null;
             }
@@ -82,9 +70,7 @@ namespace AmberEggApi.Integration.Tests.Factories
             // Assert
             var apiResponse = JsonConvert.DeserializeObject<PersonaViewModel>(await response.Content.ReadAsStringAsync());
 
-            response.StatusCode.Should().Be((int)HttpStatusCode.Created);
-            //apiResponse.IsSuccessRequest.Should().BeTrue();
-            //apiResponse.Message.Should().Be(HttpStatusCode.Created.ToString());
+            response.StatusCode.Should().Be(HttpStatusCode.Created);
 
             return apiResponse;
         }
@@ -100,14 +86,10 @@ namespace AmberEggApi.Integration.Tests.Factories
 
             // Act
             var response = await client.PutAsync($"{url}/{viewModel.Id}", requestBody);
-            // var apiResponse = JsonConvert.DeserializeObject<ApiResponse>(await response.Content.ReadAsStringAsync());
-            // var viewModelResponse =
-            //     JsonConvert.DeserializeObject<PersonaViewModel>(apiResponse.Result.ToString());
-
             var viewModelResponse = JsonConvert.DeserializeObject<PersonaViewModel>(await response.Content.ReadAsStringAsync());
+
             // Assert
             response.StatusCode.Should().Be(HttpStatusCode.OK);
-            //apiResponse.StatusCode.Should().Be((int)HttpStatusCode.OK);
             viewModelResponse.Should().BeOfType<PersonaViewModel>();
 
             viewModelResponse.Id.Should().NotBeEmpty();
